pyproject.toml
[project]
name = "testimonial-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.1",
    "flask-bcrypt>=1.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.1.0",
]
